refactor: dayjs依存を関数内に閉じ込める
