+++ src/Article.ts
@@ -1,12 +1,12 @@
 import { Tags } from './Tags';
-import dayjs from './lib/dayjs';
+import { formatYMDHm, fromNow } from './lib/dayjs';
 import { Post } from './post';
 
 export const Article = (post: Post): string => `
   <article>
     <header class="bl_text_header">
       <time class="bl_text_date">
-        ${dayjs(post.createdAt).format('YYYY-MM-DD HH:mm')}
+        ${formatYMDHm(post.createdAt)}
       </time>
     </header>
     <div class="bl_text">
@@ -15,7 +15,7 @@ export const Article = (post: Post): string => `
     </div>
     <footer class="bl_text_footer">
       <span class="bl_posts_dateago">
-        ${dayjs(post.createdAt).fromNow()}
+        ${fromNow(post.createdAt)}
       </span>
       <ul class="bl_tags">
         ${Tags(post.tags)}
