+++ src/works/citation-date.ts
@@ -1,4 +1,4 @@
-import { isNew } from '../lib/utils';
+import { isNew } from '../lib/dayjs';
 import { Citation } from './citation';
 
 const newDays = 30;
